+++ src/gradientbang/subagents/bus/local/__init__.py
@@ -1,5 +1,0 @@
-"""Local (in-process) bus implementations."""
-
-from gradientbang.subagents.bus.local.async_queue import AsyncQueueBus
-
-__all__ = ["AsyncQueueBus"]